file_managers: defaults for Mac
